% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makl_test.R
\name{makl_test}
\alias{makl_test}
\title{Test the Multiple Approximate Kernel Learning (MAKL) Model}
\usage{
makl_test(X, y, makl_model)
}
\arguments{
\item{X}{test dataset, matrix of size T x d.}

\item{y}{response vector of length T, containing only -1 and 1.}

\item{makl_model}{a list containing the MAKL model returning from makl_train().}
}
\value{
a list containing the predictions for test instances and the area under the ROC curve (AUROC) values with corresponding number of used kernels for prediction.
}
\description{
Binary classification of the test data, using the MAKL model resulted from makl_train().
}
