% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makl_test.R
\name{makl_test}
\alias{makl_test}
\title{Function for Binary Classification using a Multiple Approximate Kernel Learning (MAKL) Model}
\usage{
makl_test(X, y, makl_model)
}
\arguments{
\item{X}{data matrix of size T x d, containing the test instances.}

\item{y}{response vector of length T, containing only -1 and 1.}

\item{makl_model}{a list containing the MAKL model returning from makl_train() function.}
}
\value{
a list containing the predictions for test instances and the area under the ROC curve (auroc) values with corresponding number of used kernels for prediction.
}
\description{
Function for Binary Classification using a Multiple Approximate Kernel Learning (MAKL) Model
}
\details{
Classifies the test data, using the MAKL model resulted from makl_train().
}
